// src/dto/qa-cert-event.dto.ts
export class QaCertEventDTO {
  monLOCIdentifier: string;

  monSystemIdentifier: string;

  componentIdentifier: string;

  qaCertEventCode: string;

  qaCertEventDate: string;

  qaCertEventHour: number;

  requiredTestCode: string;

  conditionalDataBeginDate: string;

  conditionalDataBeginHour: number;

  lastTestCompletedDate: string;

  lastTestCompletedHour: number;

  lastUpdated: string;

  updatedStatusFLG: string;

  needsEvalFLG: string;

  chkSessionIdentifier: string;

  userid: string;

  addDate: string;

  updateDate: string;

  qaCertEventIdentifier: string;

  submissionIdentifier: number;

  submissionAvailabilityCode: string;

  pendingStatusCode: string;

  evalStatusCode: string;
}